refactor(o) rename to options
