+++ init.lua
@@ -1,7 +1,8 @@
 vim.cmd('ru! innit.vim')
 
-require('o')
 require('colo')
+require('options')
 require('plugins')
 
 if vim.g.neovide then require('neovide') end
+
